portal_backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.43
asyncpg==0.30.0
alembic==1.16.5
redis==6.4.0
pydantic-settings==2.11.0
python-dotenv==1.1.1
httpx==0.28.1
tenacity==9.1.2
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
email-validator==2.3.0
pytest==8.4.2
pytest-asyncio==1.2.0
aiosmtplib==3.0.2
ruff==0.13.2
